+++ rust/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "theatre_bill"
+name = "theatrical_players"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,4 +7,5 @@ edition = "2021"
 
 [dependencies]
 currency_rs = "1.3.0"
+insta = "1.35.1"
 serde_json = "1.0.114"
