+++ rust/README.md
@@ -1,0 +1,13 @@
+## Theatrical-Players-Refactoring-Kata (Rust)
+
+See the [top level readme](https://github.com/emilybache/Theatrical-Players-Refactoring-Kata) for general information about this exercise.
+
+### Installation
+
+* Install [rust](https://www.rust-lang.org/learn/get-started)
+* run `cargo test` in this directory to run tests
+
+### Approval Tests
+
+this uses [insta](https://insta.rs/) for verification tests. 
+install [cargo-insta](https://insta.rs/docs/cli/) and run `cargo insta review` after running tests to verify results.
